Move database-dependent indication review tests to integration suite.
Relocate voter and match suggestion tests that use ResetDatabase out of the unit test suite.
